- crash fix in hunspell - warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@3493 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SpellChecker/src/RichEdit.cpp
@@ -150,7 +150,7 @@ int RichEdit::GetLineCount() const
 
 void RichEdit::GetLine(int line, TCHAR *text, size_t text_len) const
 {
-	*((WORD*)text) = text_len - 1;
+	*((WORD*)text) = WORD(text_len - 1);
 	unsigned size = (unsigned) SendMessage(EM_GETLINE, (WPARAM) line, (LPARAM) text);
 	// Sometimes it likes to return size = lineLen+1, adding an \n at the end, so we remove it here
 	// to make both implementations return same size
